+++ server/src/compiler/utilities.ts
@@ -1845,9 +1845,10 @@ export function isExpressionNode(node: Node): boolean {
         case SyntaxKind.PostfixUnaryExpression:
         case SyntaxKind.BinaryExpression:
         case SyntaxKind.ConditionalExpression:
-        case SyntaxKind.SpreadElement: 
-        case SyntaxKind.ByRefElement:       
-            return true;        
+        case SyntaxKind.TemplateExpression:
+        case SyntaxKind.SpreadElement:
+        case SyntaxKind.ByRefElement:
+            return true;
         case SyntaxKind.QualifiedName:
             while (node.parent.kind === SyntaxKind.QualifiedName) {
                 node = node.parent;
@@ -4465,7 +4466,7 @@ export function createDiagnosticForNodeArray(sourceFile: SourceFileBase, nodes: 
 /**
  *  Groups of supported extensions in order of file resolution precedence. (eg, TS > TSX > DTS and seperately, CTS > DCTS)
  */
-const supportedLPCExtensions: readonly Extension[][] = [[Extension.Lpc, Extension.H, Extension.Lpc]];
+const supportedLPCExtensions: readonly Extension[][] = [[Extension.C, Extension.H, Extension.Lpc]];
 /** @internal */
 export const supportedTSExtensionsFlat: readonly Extension[] = flatten(supportedLPCExtensions);